refactor: 移除ClassicLiftCtrl并更新导入路径
