+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/controllers/DcAutoLiftCtrl.java
@@ -1,11 +1,10 @@
-package org.firstinspires.ftc.opmodes.autonomous.structure;
+package org.firstinspires.ftc.teamcode.controllers;
 
 import androidx.annotation.NonNull;
 
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 
 import org.firstinspires.ftc.teamcode.HardwareDatabase;
-import org.firstinspires.ftc.teamcode.controllers.LiftControllers;
 
 public class DcAutoLiftCtrl extends LiftControllers.DcLiftCtrl {
 	public DcAutoLiftCtrl(@NonNull final DcMotorEx target, final long targetPosition) {
